refactoring scheduling, debugging, controlling,... introduce plugin framework, cleanup...
- new bmx help with -v / -h / -H / -x / -X

- new configuration features
	- generalize resetting of configured functions using "-", e.g.: -co- or -c dev -ath0
	--test to test remaining args without applying them
	--runtime-dir /var/run/bmx to store unix-socket and pid files
	--loop_mode / -l , make batch_mode default, -b deprecated, introduce -l for loop instead
	--throw ( was --no-tunnel-rule )
	--originators, --links, --hnas, --interfaces
	--purge_timeout
	--dad_timeout
	--dev=ath0 and --dev=-ath0 (dynamically add and remove interfaces)
	-X show BatMan-eXperimental help, short description and configurable values
	--plugin bmx_plugin.so to load extended plugin functionality
	--plugin bmx_plugin.so -X or -cX to show available options including extended plugin functionality

- introduce section-like options
	--service=1.2.3.4:5 /seqno=1
	--dev=eth0 /ttl=4
	--unicast_hna=dhcp-clients /network=1.2.3.0 /netmask=24
	--throw=my-vpn /network=192.168.178.0 /netmask=24

- allow loopback interfaces as primary interfaces, e.g.: --dev lo:bmx

- new compile-time defines:
	-DNOTRAILER						- ~3K
	-DNODEBUGALL						- ~13k
	-DNOTUNNEL  	(only affects this node)		- ~23k
	-DNOSRV  	(only affects this node)		- ~3k
	-DNOVIS  	(only affects this node)		- ~2k
	-DNODEPRECATED	(for backward compatibility)		- ~2k

- new advanced compile-time defines (please dont touch):
	-DNOHNA		(affects all nodes in network)		- ~6k
	-DNOPARANOIA	(makes bug-hunting impossible)		- ~2k
	-DEXTDEBUG	(this eats your cpu)			+ ~0k
	-DTESTDEBUG	(testing syntax of __VA_ARGS__ dbg...() macros)
	-DWITHUNUSED	(includes yet unused stuff)

- allow full-broadcast destination addresses to facilitate IP renumbering

- track configured ip-rules and ip-routes and dont configure things twice

- better conform to common coding-styles: -pedantic -Wall -W -Wno-unused-parameter -std=gnu99

- check for duplicate ip configuration on local node

- remove lots of global variables

- plugin framework allowing registration of:
	- parameters
	- data-structures
	- scheduled functions
	- ...

- support muting of selected debug messages

- new bit-array processing funtions

- remove usage of libpthread

- use ip address of primary interface as preferred src for all interfaces

- periodically check for changed ip_forward and other kernel configs

+++ hash.h
@@ -18,8 +18,8 @@
  */
 
 
-typedef int (*hashdata_compare_cb)(void *, void *, int);
-typedef int (*hashdata_choose_cb)(void *, int, int);
+typedef int (*hashdata_compare_cb)(void *, void *, size_t);
+typedef int (*hashdata_choose_cb)(void *, int, size_t);
 typedef void (*hashdata_free_cb)(void *);
 
 struct element_t {
@@ -47,8 +47,8 @@ struct hashtable_t {
 					 * and the size the second */
 };
 
-int compare_key( void *data1, void *data2, int key_size );
-int choose_key( void *data, int size, int key_size );
+int compare_key( void *data1, void *data2, size_t key_size );
+int choose_key( void *data, int size, size_t key_size );
 
 
 /* clears the hash */
@@ -86,7 +86,9 @@ void 				*hash_find(struct hashtable_t *hash, void *keydata);
 struct hashtable_t	*hash_resize(struct hashtable_t *hash, int size);
 
 /* print the hash table for debugging */
+/*
 void 				 hash_debug( struct hashtable_t *hash);
+*/
 
 /* iterate though the hash. first element is selected with iter_in NULL.
  * use the returned iterator to access the elements until hash_it_t returns NULL. */
